requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
pillow==10.4.0
pydantic==2.9.2
pytest==8.3.3
httpx==0.27.2